requirements-pip.txt
numpy>=1.13.3
scipy>=1.0.0
tensorflow-gpu>=1.6.0
moviepy>=0.2.3.2
Pillow>=3.1.1
lmdb>=0.93
opencv-python>=3.4.0.12
cryptography>=2.1.4
h5py>=2.7.1
six>=1.11.0
